+++ src/scene_server/admin_server/imports.go
@@ -96,5 +96,5 @@ import (
 
 	// v3.12.x
 	_ "configcenter/src/scene_server/admin_server/upgrader/y3.12.202309211735"
-	_ "configcenter/src/scene_server/admin_server/upgrader/y3.12.202310101010"
+	_ "configcenter/src/scene_server/admin_server/upgrader/y3.12.202310301633"
 )
