+++ src/scene_server/admin_server/upgrader/y3.12.202310301633/add_recorded_fields.go
@@ -15,7 +15,7 @@
  * to the current version of the project delivered to anyone in the future.
  */
 
-package y3_12_202310101010
+package y3_12_202310301633
 
 import (
 	"context"
