+++ src/source_controller/coreservice/core/instances/instance.go
@@ -499,7 +499,7 @@ func (m *instanceManager) SearchModelInstance(kit *rest.Kit, objID string, input
 	// parse vip fields for processes
 	fields, vipFields := hooks.ParseVIPFieldsForProcessHook(inputParam.Fields, tableName)
 
-	fields, status, err := util.GetFieldStatus(fields)
+	fields, status, err := GetFieldStatus(fields)
 	if err != nil {
 		blog.Errorf("get filed status failed, fields:%v, err: %v, rid: %s", fields, err, kit.Rid)
 		return nil, err
@@ -509,7 +509,7 @@ func (m *instanceManager) SearchModelInstance(kit *rest.Kit, objID string, input
 }
 
 func (m *instanceManager) searchModelInstance(kit *rest.Kit, objID string, inputParam metadata.QueryCondition,
-	fields []string, vipFields []string, status *util.FieldStatus) (*metadata.QueryResult, error) {
+	fields []string, vipFields []string, status *FieldStatus) (*metadata.QueryResult, error) {
 
 	tableName := common.GetInstTableName(objID, kit.SupplierAccount)
 	instItems := make([]mapstr.MapStr, 0)
@@ -706,4 +706,40 @@ func (m *instanceManager) CascadeDeleteModelInstance(kit *rest.Kit, objID string
 	}
 
 	return &metadata.DeletedCount{Count: uint64(len(origins))}, nil
+}
+
+// FieldStatus model field status
+type FieldStatus struct {
+	ExistCreateAt   bool
+	ExistCreateTime bool
+	ExistUpdateAt   bool
+	ExistLastTime   bool
+}
+
+// GetFieldStatus get field status
+func GetFieldStatus(fields []string) ([]string, *FieldStatus, error) {
+	status := &FieldStatus{ExistCreateAt: true, ExistCreateTime: true, ExistUpdateAt: true, ExistLastTime: true}
+	if len(fields) == 0 {
+		return fields, status, nil
+	}
+
+	// 旧数据用create_time和last_time分别记录了实例的创建和更新时间，如果bk_created_at和bk_updated_at字段没值，需要把旧值赋过来
+	fieldMap := make(map[string]struct{})
+	for _, field := range fields {
+		fieldMap[field] = struct{}{}
+	}
+
+	_, status.ExistCreateAt = fieldMap[common.BKCreatedAt]
+	_, status.ExistCreateTime = fieldMap[common.CreateTimeField]
+	if status.ExistCreateAt && !status.ExistCreateTime {
+		fields = append(fields, common.CreateTimeField)
+	}
+
+	_, status.ExistUpdateAt = fieldMap[common.BKUpdatedAt]
+	_, status.ExistLastTime = fieldMap[common.LastTimeField]
+	if status.ExistUpdateAt && !status.ExistLastTime {
+		fields = append(fields, common.LastTimeField)
+	}
+
+	return fields, status, nil
 }
